Char type, new sum implementation, & more
New Char type for making lists/Arrays of Char to emulate a string, but with the ability to treat it like any other sequence. The new "chars" function provides a friendly interface for converting strings to Arrays of Char.

Arrays also provide a nice representation of [Char] now.

Arrays are now tuples under the hood, instead of lists. Representation is still the same.

Also added a "sum" function that does the same thing as the standard "sum" function, but it also works on things besides numbers.